+++ library/src/main/java/com/evernote/android/job/DailyJob.java
@@ -8,9 +8,9 @@ import android.support.annotation.WorkerThread;
 import com.evernote.android.job.util.JobCat;
 import com.evernote.android.job.util.support.PersistableBundleCompat;
 
-import net.vrallev.android.cat.CatLog;
-
 import java.util.Calendar;
+import java.util.HashSet;
+import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
 /**
@@ -19,23 +19,25 @@ import java.util.concurrent.TimeUnit;
  *
  * @author rwondratschek
  */
-@SuppressWarnings("unused")
+@SuppressWarnings({"unused", "WeakerAccess"})
 public abstract class DailyJob extends Job {
 
-    private static final CatLog CAT = new JobCat("DailyJob");
+    private static final JobCat CAT = new JobCat("DailyJob");
 
     @VisibleForTesting
     /*package*/ static final String EXTRA_START_MS = "EXTRA_START_MS";
     @VisibleForTesting
     /*package*/ static final String EXTRA_END_MS = "EXTRA_END_MS";
+    @VisibleForTesting
+    private static final String EXTRA_ONCE = "EXTRA_ONCE";
 
     private static final long DAY = TimeUnit.DAYS.toMillis(1);
 
     /**
      * Schedules your daily job. A builder is required for this method call. Within the builder, you may specify
      * additional requirements and/or extras for the job. However, a daily job may not be exact,
-     * periodic or transient. Since the rescheduling of a daily job when requirements aren't met
-     * (e.g. low internet connectivity) isn't useful, the enforcing of such requirements isn't supported either.
+     * periodic or transient. If the requirements are enforced but are not met when the job runs, then the
+     * daily job will only be rescheduled for the next day. The back-off criteria is ignored in this case.
      * <br>
      * <br>
      * Daily jobs should use a unique tag and their classes shouldn't be reused for other jobs.
@@ -62,6 +64,26 @@ public abstract class DailyJob extends Job {
      */
     public static int schedule(@NonNull JobRequest.Builder baseBuilder, long startMs, long endMs) {
         return schedule(baseBuilder, true, startMs, endMs);
+    }
+
+    /**
+     * Schedules the daily job only once and runs it immediately. This is helpful if you want to reuse your job
+     * and want to trigger the execution immediately. It's possible to schedule a daily job normally with
+     * {@link #schedule(JobRequest.Builder, long, long)} and this method at the same time to trigger the
+     * execution immediately.
+     *
+     * @param baseBuilder The builder of your daily job.
+     * @return The unique ID for this job.
+     */
+    public static int startNowOnce(@NonNull JobRequest.Builder baseBuilder) {
+        PersistableBundleCompat extras = new PersistableBundleCompat();
+        extras.putBoolean(EXTRA_ONCE, true);
+
+        return baseBuilder
+                .startNow()
+                .addExtras(extras)
+                .build()
+                .schedule();
     }
 
     private static int schedule(@NonNull JobRequest.Builder builder, boolean newJob, long startMs, long endMs) {
@@ -98,16 +120,23 @@ public abstract class DailyJob extends Job {
 
         builder.addExtras(extras);
 
+        if (newJob) {
+            // cancel all previous jobs, but not the one scheduled immediately
+            JobManager manager = JobManager.instance();
+            Set<JobRequest> requests = new HashSet<>(manager.getAllJobRequestsForTag(builder.mTag));
+            for (JobRequest request : requests) {
+                if (!request.isExact() || request.getStartMs() != JobRequest.START_NOW) {
+                    manager.cancel(request.getJobId());
+                }
+            }
+        }
+
         JobRequest request = builder
                 .setExecutionWindow(Math.max(1L, startDelay), Math.max(1L, endDelay))
-                .setUpdateCurrent(newJob)
                 .build();
 
         if (newJob && (request.isExact() || request.isPeriodic() || request.isTransient())) {
             throw new IllegalArgumentException("Daily jobs cannot be exact, periodic or transient");
-        }
-        if (newJob && (request.requirementsEnforced())) {
-            throw new IllegalArgumentException("Daily jobs cannot enforce requirements");
         }
 
         return request.schedule();
@@ -115,9 +144,11 @@ public abstract class DailyJob extends Job {
 
     @NonNull
     @Override
-    protected final Result onRunJob(Params params) {
+    protected final Result onRunJob(@NonNull Params params) {
         PersistableBundleCompat extras = params.getExtras();
-        if (!extras.containsKey(EXTRA_START_MS) || !extras.containsKey(EXTRA_END_MS)) {
+        boolean runOnce = extras.getBoolean(EXTRA_ONCE, false);
+
+        if (!runOnce && (!extras.containsKey(EXTRA_START_MS) || !extras.containsKey(EXTRA_END_MS))) {
             CAT.e("Daily job doesn't contain start and end time");
             return Result.FAILURE;
         }
@@ -125,7 +156,13 @@ public abstract class DailyJob extends Job {
         DailyJobResult result = null;
 
         try {
-            result = onRunDailyJob(params);
+            if (meetsRequirements()) {
+                result = onRunDailyJob(params);
+            } else {
+                result = DailyJobResult.SUCCESS; // reschedule
+                CAT.i("Daily job requirements not met, reschedule for the next day");
+            }
+
         } finally {
             if (result == null) {
                 // shouldn't happen if the job follows the contract
@@ -133,16 +170,18 @@ public abstract class DailyJob extends Job {
                 CAT.e("Daily job result was null");
             }
 
-            JobRequest request = params.getRequest();
-            if (result == DailyJobResult.SUCCESS) {
-                CAT.i("Rescheduling daily job %s", request);
+            if (!runOnce) {
+                JobRequest request = params.getRequest();
+                if (result == DailyJobResult.SUCCESS) {
+                    CAT.i("Rescheduling daily job %s", request);
 
-                // don't update current, it would cancel this currently running job
-                schedule(request.createBuilder(), false,
-                        extras.getLong(EXTRA_START_MS, 0) % DAY, extras.getLong(EXTRA_END_MS, 0L) % DAY);
+                    // don't update current, it would cancel this currently running job
+                    schedule(request.createBuilder(), false,
+                            extras.getLong(EXTRA_START_MS, 0) % DAY, extras.getLong(EXTRA_END_MS, 0L) % DAY);
 
-            } else {
-                CAT.i("Cancel daily job %s", request);
+                } else {
+                    CAT.i("Cancel daily job %s", request);
+                }
             }
         }
 
@@ -172,7 +211,7 @@ public abstract class DailyJob extends Job {
     @SuppressWarnings("WeakerAccess")
     @NonNull
     @WorkerThread
-    protected abstract DailyJobResult onRunDailyJob(Params params);
+    protected abstract DailyJobResult onRunDailyJob(@NonNull Params params);
 
     public enum DailyJobResult {
         /**