+++ library/src/main/java/com/evernote/android/job/Job.java
@@ -36,8 +36,6 @@ import com.evernote.android.job.util.Device;
 import com.evernote.android.job.util.JobCat;
 import com.evernote.android.job.util.support.PersistableBundleCompat;
 
-import net.vrallev.android.cat.CatLog;
-
 import java.lang.ref.WeakReference;
 
 /**
@@ -48,7 +46,7 @@ import java.lang.ref.WeakReference;
 @SuppressWarnings({"unused", "WeakerAccess"})
 public abstract class Job {
 
-    private static final CatLog CAT = new JobCat("Job");
+    private static final JobCat CAT = new JobCat("Job");
 
     public enum Result {
         /**
@@ -113,11 +111,12 @@ public abstract class Job {
      */
     @NonNull
     @WorkerThread
-    protected abstract Result onRunJob(Params params);
+    protected abstract Result onRunJob(@NonNull Params params);
 
     /*package*/ final Result runJob() {
         try {
-            if (meetsRequirements()) {
+            // daily jobs check the requirements manually
+            if (this instanceof DailyJob || meetsRequirements()) {
                 mResult = onRunJob(getParams());
             } else {
                 mResult = getParams().isPeriodic() ? Result.FAILURE : Result.RESCHEDULE;
@@ -143,7 +142,7 @@ public abstract class Job {
         // override me
     }
 
-    private boolean meetsRequirements() {
+    /*package*/ boolean meetsRequirements() {
         if (!getParams().getRequest().requirementsEnforced()) {
             return true;
         }