+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-## 1.2.1
+## 1.2.1 (2017-11-23)
 * Add option to override the default background executor, see #292
 * Don't keep a strong reference to finished jobs to allow freeing memory, see #299
+* Allow running a daily job immediately once, this is helpful for testing purposes, see #317
+* Allow enforcing requirements for daily jobs, see #313
+* Remove the Cat dependency for logging, see 295
+* Add `@NonNull` annotation to `param` parameter, see #321 (thanks for the contribution @Jawnnypoo)
 
 ## 1.2.0 (2017-10-05)